Convert local-storage hook from js -> ts
